split up files, force exit
- always use io.exit for exit (sometimes indirectly)
